Updates in implementation and some fixes.

+++ src/main/java/stories/profilesedition/ProfilesEditionController.java
@@ -9,6 +9,8 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.cell.TextFieldTableCell;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
+import pojos.kf2factory.Kf2Common;
+import pojos.kf2factory.Kf2Factory;
 import pojos.session.Session;
 import utils.Utils;
 
@@ -42,10 +44,18 @@ public class ProfilesEditionController implements Initializable {
 
     @FXML
     private void addProfileOnAction() {
-        Optional<String> profileNameOpt = Utils.OneTextInputDialog("Add a new profile", "Enter profile name:");
         try {
+            String installationFolder = facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER);
+            if (StringUtils.isBlank(installationFolder)) {
+                Utils.warningDialog("You can not add a new profile!","You need to define an installation folder in Install/Update section.");
+                return;
+            }
+            Optional<String> profileNameOpt = Utils.OneTextInputDialog("Add a new profile", "Enter profile name:");
             if (profileNameOpt.isPresent() && StringUtils.isNotBlank(profileNameOpt.get())){
-                profilesTable.getItems().add(facade.createNewProfile(profileNameOpt.get()));
+                String profileName = profileNameOpt.get().replaceAll(" ", "_");
+                profilesTable.getItems().add(facade.createNewProfile(profileName));
+                Kf2Common kf2Common = Kf2Factory.getInstance();
+                kf2Common.createConfigFolder(installationFolder, profileName);
             }
         } catch (Exception e) {
             Utils.errorDialog("The profile can not be created!", "See stacktrace for more details", e);
@@ -55,12 +65,17 @@ public class ProfilesEditionController implements Initializable {
     @FXML
     private void removeProfileOnAction() {
         try {
+            String installationFolder = facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER);
+            if (StringUtils.isBlank(installationFolder)) {
+                Utils.warningDialog("You can not remove a profile!","You need to define an installation folder in Install/Update section.");
+                return;
+            }
             int selectedIndex = profilesTable.getSelectionModel().getSelectedIndex();
             if (selectedIndex >= 0) {
                 ProfileDto selectedProfile = profilesTable.getSelectionModel().getSelectedItem();
                 if (facade.deleteSelectedProfile(selectedProfile.getName())) {
                     profilesTable.getItems().remove(selectedIndex);
-                    File profileConfigFolder = new File(facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER) + "\\KFGame\\Config\\" + Session.getInstance().getActualProfile().getName());
+                    File profileConfigFolder = new File(facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER) + "/KFGame/Config/" + selectedProfile.getName());
                     FileUtils.deleteDirectory(profileConfigFolder);
                 } else {
                     Utils.errorDialog("The profile can not be deleted from database", "Delete operation is aborted!", null);
@@ -77,13 +92,18 @@ public class ProfilesEditionController implements Initializable {
     private void profileNameColumnOnEditCommit(TableColumn.CellEditEvent<?,?> event) {
         int edittedRowIndex = event.getTablePosition().getRow();
         String oldProfileName = (String)event.getOldValue();
-        String newProfileName = (String)event.getNewValue();
+        String newProfileName = ((String)event.getNewValue()).replaceAll(" ", "_");
         try {
+            String installationFolder = facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER);
+            if (StringUtils.isBlank(installationFolder)) {
+                Utils.warningDialog("You can not edit a profile!","You need to define an installation folder in Install/Update section.");
+                return;
+            }
             ProfileDto updatedProfileDto = facade.updateChangedProfile(oldProfileName, newProfileName);
             if (updatedProfileDto != null) {
                 profilesTable.getItems().remove(edittedRowIndex);
                 profilesTable.getItems().add(updatedProfileDto);
-                String configFolder = facade.findPropertyValue(Constants.KEY_INSTALLATION_FOLDER) + "\\KFGame\\Config\\";
+                String configFolder = installationFolder + "/KFGame/Config/";
                 File oldProfileConfigFolder = new File(configFolder + oldProfileName);
                 if (oldProfileConfigFolder.exists() && oldProfileConfigFolder.isDirectory()) {
                     File newProfileConfigFolder = new File(configFolder + newProfileName);
